requirements.txt
Django==4
psycopg2-binary
psycopg2-binary==2.9.9

djangorestframework~=3.14.0
model-bakery==1.17.0
